refactor: move slice tests
